[sw/crypto] Move ibex driver back from base lib into cryptolib
