+++ sw/device/lib/crypto/drivers/ibex.h
@@ -1,8 +1,8 @@
 // Copyright lowRISC contributors (OpenTitan project).
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
-#ifndef OPENTITAN_SW_DEVICE_LIB_BASE_IBEX_H_
-#define OPENTITAN_SW_DEVICE_LIB_BASE_IBEX_H_
+#ifndef OPENTITAN_SW_DEVICE_LIB_CRYPTO_DRIVERS_IBEX_H_
+#define OPENTITAN_SW_DEVICE_LIB_CRYPTO_DRIVERS_IBEX_H_
 
 #include <stdbool.h>
 #include <stddef.h>
@@ -50,4 +50,4 @@ status_t ibex_rnd_status_read(uint32_t *rnd_status);
  */
 uint32_t ibex_rnd_uint32(void);
 
-#endif  // OPENTITAN_SW_DEVICE_LIB_BASE_IBEX_H_
+#endif  // OPENTITAN_SW_DEVICE_LIB_CRYPTO_DRIVERS_IBEX_H_
