+++ sw/device/lib/crypto/impl/BUILD
@@ -105,8 +105,8 @@ cc_library(
         ":status",
         "//sw/device/lib/base:hardened",
         "//sw/device/lib/base:hardened_memory",
-        "//sw/device/lib/base:ibex",
         "//sw/device/lib/crypto/drivers:hmac",
+        "//sw/device/lib/crypto/drivers:ibex",
         "//sw/device/lib/crypto/drivers:kmac",
         "//sw/device/lib/crypto/include:datatypes",
     ],
@@ -248,8 +248,8 @@ cc_library(
         ":keyblob",
         "//sw/device/lib/base:hardened",
         "//sw/device/lib/base:hardened_memory",
-        "//sw/device/lib/base:ibex",
         "//sw/device/lib/crypto/drivers:hmac",
+        "//sw/device/lib/crypto/drivers:ibex",
         "//sw/device/lib/crypto/drivers:kmac",
     ],
 )
