+++ sw/device/lib/crypto/impl/mac.c
@@ -5,8 +5,8 @@
 #include "sw/device/lib/crypto/include/mac.h"
 
 #include "sw/device/lib/base/hardened_memory.h"
-#include "sw/device/lib/base/ibex.h"
 #include "sw/device/lib/crypto/drivers/hmac.h"
+#include "sw/device/lib/crypto/drivers/ibex.h"
 #include "sw/device/lib/crypto/drivers/kmac.h"
 #include "sw/device/lib/crypto/impl/integrity.h"
 #include "sw/device/lib/crypto/impl/keyblob.h"
